fix: [M2048] Key Collect status colour on stable value
The status border colour was keyed on the English status label, so it
disappeared once the label was translated. Derive a stable statusValue
in CollectRecordsMixin and colour each row from that instead.

Migrate the row styling from a styled-component to Collect.module.scss.
The colour map stays in JS and reaches the module through a CSS custom
property, keeping theme.ts as the single source of truth rather than
duplicating colours into SCSS where the two theme systems have already
drifted. Rename collectConstants to TypeScript and add regression tests.
