+++ src/components/pages/Collect/Collect.module.scss
@@ -1,0 +1,3 @@
+.record-status-row {
+  border-left: 4px solid var(--record-status-color, transparent);
+}
